Moved `saveSites` to test util * this used to be in services, but is not used in production
